Arduino and PurrData first code
Arduino code to recognise knocks.
PurrData to play audio when knock is recognised.
Wav files are also uploaded.
If going to use the PurrData, edit the code in the Purr to your own wav file location and remember to edit the port number "open 3" to the one used in the arduino program.